Expose SdkLogListener instance through setNodeLogStream
Export SdkLogListener to it's own file
